+++ lib/src/services/ds_file.service.dart
@@ -1,25 +1,25 @@
 import 'dart:io';
 
 import 'package:dio/dio.dart';
-import 'package:get/get.dart' hide Response;
 import 'package:open_filex/open_filex.dart';
 import 'package:path/path.dart' as path_utils;
 import 'package:path_provider/path_provider.dart';
 import 'package:url_launcher/url_launcher_string.dart';
 
-import '../themes/colors/ds_colors.theme.dart';
-
 abstract class DSFileService {
   static Future<void> open(
-    final String fileName,
+    final String filename,
     final String url, {
-    void Function(bool)? onDownloadStateChange,
+    final void Function(bool)? onDownloadStateChange,
+    final Map<String, String?>? httpHeaders,
   }) async {
     final path = await download(
       url,
-      fileName,
+      filename,
       onDownloadStateChange: onDownloadStateChange,
+      httpHeaders: httpHeaders,
     );
+
     if (path?.isEmpty ?? true) {
       return;
     }
@@ -30,47 +30,45 @@ abstract class DSFileService {
       case ResultType.done:
         break;
       case ResultType.noAppToOpen:
-        launchUrlString(url, mode: LaunchMode.externalApplication);
+        launchUrlString(
+          url,
+          mode: LaunchMode.externalApplication,
+        );
         break;
       default:
-        _showAlert('Erro', 'Visualização indisponível');
+        throw Exception(result.message);
     }
   }
 
   static Future<String?> download(
     final String url,
-    final String fileName, {
+    final String filename, {
     final String? path,
-    void Function(bool)? onDownloadStateChange,
+    final void Function(bool)? onDownloadStateChange,
+    final Map<String, String?>? httpHeaders,
   }) async {
     try {
       onDownloadStateChange?.call(true);
       final savedFilePath = path_utils.join(
-          path ?? (await getTemporaryDirectory()).path, fileName);
+          path ?? (await getTemporaryDirectory()).path, filename);
 
       if (await File(savedFilePath).exists()) {
         return savedFilePath;
       }
 
-      final response = await Dio().download(url, savedFilePath);
+      final response = await Dio().download(
+        url,
+        savedFilePath,
+        options: Options(
+          headers: httpHeaders,
+        ),
+      );
 
       if (response.statusCode == 200) return savedFilePath;
-    } catch (e) {
-      _showAlert('Erro', 'Download indisponível');
     } finally {
       onDownloadStateChange?.call(false);
     }
 
     return null;
-  }
-
-  //TODO: Use the DS's widget when avalible and question the PD about it
-  static void _showAlert(
-    final String title,
-    final String message,
-  ) {
-    Get.snackbar(title, message,
-        snackPosition: SnackPosition.BOTTOM,
-        backgroundColor: DSColors.neutralLightSnow);
   }
 }