+++ lib/src/utils/ds_utils.util.dart
@@ -3,11 +3,9 @@ import '../models/ds_country.model.dart';
 /// All utility constants that are used by this Design System.
 abstract class DSUtils {
   static const packageName = 'blip_ds';
-  static const defaultAnimationDuration = Duration(
-    milliseconds: 300,
-  );
   static const bubbleMinSize = 240.0;
   static const bubbleMaxSize = 480.0;
+  static const defaultAnimationDuration = Duration(milliseconds: 300);
 
   static const countriesList = <DSCountry>[
     DSCountry(